+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 FIXES:
 
 1. Fix SEC-SIGLOG payload definition - thanks to @Niederb for contribution.
+1. Cater for alternate (v1, v2) versions of SEC-SIG message = Fixes [#156](https://github.com/semuconsulting/pyubx2/issues/156).
 
 ### RELEASE 1.2.44